requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.22.0
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.1.0
SQLAlchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiosqlite==0.21.0
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
prometheus-fastapi-instrumentator==6.1.0

# ML dependencies for pricing optimization
catboost==1.2.8
scikit-learn==1.7.0
pandas==2.3.1
numpy==2.3.1
redis==5.0.1
pika==1.3.2
loguru==0.7.2
# pickle5==0.0.11  # Removed - compatibility issues with modern Python